requirements.txt
grpcio==1.68.0
grpcio-tools==1.68.0
pytest==7.4.0
docker==7.1.0